move TextFieldBase to the internal namespace
